refactor: update util imports
